Add black formatting

+++ app/dao/broadcast_service_dao.py
@@ -37,7 +37,7 @@ def set_broadcast_service_type(service, service_mode, broadcast_channel, provide
         # Email auth is an exception to the other service permissions (which relate to what type
         # of notifications a service can send) where a broadcast service is allowed to have the
         # email auth permission (but doesn't have to)
-        ServicePermission.permission != EMAIL_AUTH_TYPE
+        ServicePermission.permission != EMAIL_AUTH_TYPE,
     ).delete()
 
     # Refresh the service object as it has references to the service permissions but we don't yet
@@ -59,27 +59,17 @@ def set_broadcast_service_type(service, service_mode, broadcast_channel, provide
         service.go_live_at = None
 
     # Remove all user permissions apart from view_activity for the service users and invited users
-    Permission.query.filter(
-        Permission.service_id == service.id,
-        Permission.permission != VIEW_ACTIVITY
-    ).delete()
-    InvitedUser.query.filter_by(
-        service_id=service.id,
-        status=INVITE_PENDING
-    ).update({'permissions': VIEW_ACTIVITY})
+    Permission.query.filter(Permission.service_id == service.id, Permission.permission != VIEW_ACTIVITY).delete()
+    InvitedUser.query.filter_by(service_id=service.id, status=INVITE_PENDING).update({"permissions": VIEW_ACTIVITY})
 
     # Revoke any API keys to avoid a regular API key being used to send alerts
     ApiKey.query.filter_by(
         service_id=service.id,
         expiry_date=None,
-    ).update({
-        ApiKey.expiry_date: datetime.utcnow()
-    })
+    ).update({ApiKey.expiry_date: datetime.utcnow()})
 
     # Add service to organisation
-    organisation = Organisation.query.filter_by(
-        id=current_app.config['BROADCAST_ORGANISATION_ID']
-    ).one()
+    organisation = Organisation.query.filter_by(id=current_app.config["BROADCAST_ORGANISATION_ID"]).one()
     service.organisation_id = organisation.id
     service.organisation_type = organisation.organisation_type
     service.crown = organisation.crown
